fix: resolve react-refresh/only-export-components warnings
Move getDisplayLanguage utility function to a dedicated utils file and
add eslint-disable comment for FileDiffSession where hook + Provider
co-location is an intentional design pattern.

Changes:
- Create utils/irLanguage.ts with getDisplayLanguage function
- Update imports in TritonIRs.tsx, SingleCodeViewer.tsx,
  CodeComparisonView.tsx, CodeView.tsx
- Add eslint-disable for FileDiffSession.tsx (hook + Provider pattern)

Files modified: 5
New files: 1 (utils/irLanguage.ts)
